refactor project tree files - to adapt for more models type later.
