+++ EDA/output_EDA/cleaned_data_based_output/score_and_rating_distributions.png
