+++ EDA/output_EDA/games_invest_graphs/game_end_reason_distribution.png
