+++ EDA/output_EDA/games_invest_graphs/numerical_distributions/game_duration_seconds_distribution.png
