+++ EDA/output_EDA/games_invest_graphs/numerical_distributions/increment_seconds_distribution.png
