+++ EDA/output_EDA/games_invest_graphs/numerical_distributions/initial_time_seconds_distribution.png
