+++ EDA/output_EDA/games_invest_graphs/numerical_distributions/max_overtime_minutes_distribution.png
