+++ EDA/output_EDA/games_invest_graphs/rating_mode_distribution.png
