+++ EDA/output_EDA/games_invest_graphs/time_analysis/games_by_day_of_week.png
