+++ EDA/output_EDA/games_invest_graphs/time_control_distribution.png
