+++ EDA/output_EDA/train_invest_graph/avg_score_vs_avg_rating_humans.png
