+++ EDA/output_EDA/train_invest_graph/final_scores_distribution.png
