+++ EDA/output_EDA/train_invest_graph/player_rating_distribution.png
