+++ EDA/output_EDA/train_invest_graph/rating_vs_score_correlation_heatmap.png
