+++ EDA/output_EDA/train_invest_graph/score_per_rating.png
