+++ EDA/output_EDA/train_invest_graph/total_games_distribution.png
