+++ EDA/output_EDA/train_invest_graph/total_games_vs_avg_rating_no_outliers.png
