+++ EDA/output_EDA/turns_invest_graphs/advanced_invest/game_length_vs_score.png
