+++ EDA/output_EDA/turns_invest_graphs/advanced_invest/most_common_letters_rack.png
