+++ EDA/output_EDA/turns_invest_graphs/basic_invest/points_per_turn_number.png
