+++ EDA/output_EDA/turns_invest_graphs/basic_invest/score_distribution.png
