+++ EDA/output_EDA/turns_invest_graphs/basic_invest/turn_type_distribution.png
