+++ EDA/output_EDA/turns_invest_graphs/relevant_to_rank/avg_points_per_player_distribution.png
