+++ Model/Pipeline/Transformers/Data_Cleaning/drop_features_from_train.py
@@ -1,4 +1,4 @@
-from Model.Transformers.abstract_tranformer import AbstractTransformer
+from Model.Pipeline.Transformers.abstract_tranformer import AbstractTransformer
 from typing import List, Dict
