+++ Model/Pipeline/Transformers/Data_Cleaning/merge_features_from_games.py
@@ -1,4 +1,4 @@
-from Model.Transformers.abstract_tranformer import AbstractTransformer
+from Model.Pipeline.Transformers.abstract_tranformer import AbstractTransformer
 from typing import List, Dict
