+++ Model/Pipeline/Transformers/Feature_Engineering/add_mean_average_letter_score.py
@@ -1,4 +1,4 @@
-from Model.Transformers.abstract_tranformer import AbstractTransformer
+from Model.Pipeline.Transformers.abstract_tranformer import AbstractTransformer
 import pandas as pd
 from typing import Dict
