+++ Model/Pipeline/Transformers/abstract_tranformer.py
